Checkout this repo to /opt/infrastructure
At the root of the infrastructure project will be docker-compose
infrastructure. There's a couple ways to this:

All of the above have issues

1. Use NixOS setup infrastructure repo to access docker-compose.

Using NixOS is not viable as causes a chicken & egg issue for hash
pinning. A seperate repo could be used, but because of how Nix
functions, prototyping changes would be harder to do in place.

2. Use Terraform/Cloud-Init to copy over docker compose files & other dependencies

Terraform doesn't really support managing local files. We can do this
individually & piecemeal but this is rather burdensome as the
configuration grows. The local provisioner can be used instead but this
also is not fully supported [1] & recursive copies folders are more
complicated.

3. Checkout the repo from github

This also isn't ideal as this will checkout only the default branch.
Additional logic could be added to allow for more arbitrary versioning,
but complications stem from not using the local copy when terraform
runs. I don't think this is currently that big of a problem, but may add
some complications when attempting to use CI to deploy IaC changes.

Currently I'm going with option 3, though I may attempt option 2 in the future.

[1] https://stackoverflow.com/questions/74392871/

We'll want to modify the docker-compose for prototyping changes in
place. There was a couple ways to do this

Adding this to the /etc/NIXOS_LUSTRATE file will ensure that it will
survive the nixos-infect.