Adding POC for docker-compose image
This deploys the files located in the ./docker directory to
/opt/infrastructure on the new machine. This was done to allow
modifications to the docker configuration without a full rebuild.

This is done by archiving the ./docker directory using the
hachicorp/archive provider. The archive is then is base64 encoded
included in the cloud-init. The archives are deterministic on the same
machine but may not be so for other systems.